refactor(import): unify query structure and stabilize import jobs
- move shared SPARQL helpers to queries/_shared.js
- fix war.job per-category error handling
- remove invalid medicine task placeholder
- add dotenv loading for CLI import jobs
- verify war + medicine imports with dry-run
